refactor: Move SubCommand class into Command.ts file, see explaination in the file.
