+++ src/classes/CommandHandlerError.ts
