+++ src/classes/commands/SubCommandContext.ts
@@ -1,5 +1,5 @@
 import {CommandContext, CommandContextBuilder} from './CommandContext.js';
-import {SubCommand} from './SubCommand.js';
+import {SubCommand} from './Command.js';
 
 export interface SubCommandContextBuilder extends CommandContextBuilder {
 	subCommand: SubCommand;
